init addmentoring graphl
